+++ kristine.h
@@ -1,5 +1,5 @@
 #ifndef KRISTINE.H
 #define KRISTINE.H
-void semConRegistrationFees(double totalExpenses);
-void hotelExpenses(int tripDays, double totalExpenses, double returnExpenses, double savedExpenses, double allowableExpenses);
+void semConRegistrationFees(float totalExpenses);
+void hotelExpenses(int tripDays, float totalExpenses, float returnExpenses, float savedExpenses, float allowableExpenses);
 #endif
